finished the refactor.  All files in routes are now organized
