+++ api/routes/affiliations/openAPI.js
@@ -43,43 +43,6 @@ const updated_at = {
   description: 'Timestamp of last update'
 };
 
-const filter_status = {
-  type: 'string',
-  description: 'The filter status of the affiliations for this US State',
-  enum: ['pending', 'active', 'inactive']
-};
-
-const stateIdParameter = {
-  name: 'stateId',
-  in: 'path',
-  description: state_id.description,
-  required: true,
-  schema: {
-    type: state_id.type
-  }
-};
-
-const filterStatusParameter = {
-  in: 'query',
-  name: 'status',
-  description: filter_status.description,
-  required: false,
-  schema: {
-    type: filter_status.type,
-    enum: filter_status.enum
-  }
-};
-
-const idParameter = {
-  name: 'id',
-  in: 'path',
-  description: id.description,
-  required: true,
-  schema: {
-    type: id.type
-  }
-};
-
 const affiliationSchema = {
   type: 'object',
   properties: {
@@ -94,28 +57,6 @@ const affiliationSchema = {
 };
 
 const tags = ['Affiliations'];
-
-const postAffiliations = {
-  post: {
-    tags,
-    description:
-      'Create a request for the currently logged in user to be affiliated with a US State',
-    parameters: [stateIdParameter],
-    responses: {
-      201: {
-        description: 'Record was created',
-        content: jsonResponse(id)
-      },
-      400: {
-        description: 'Record does not exist or US State ID is invalid'
-      },
-      401: {
-        ...responses.unauthed[401]
-      }
-    },
-    security: [{ bearerAuth: [] }]
-  }
-};
 
 const getMyAffiliations = {
   get: {
@@ -132,36 +73,6 @@ const getMyAffiliations = {
     security: [{ bearerAuth: [] }]
   }
 };
-
-const patchAffiliation = {
-  patch: {
-    tags,
-    description: 'Update a user affiliation with a US State',
-    parameters: [stateIdParameter, idParameter],
-    requestBody: {
-      required: true,
-      content: jsonResponse({
-        type: 'object',
-        properties: {
-          roleId: role_id,
-          status
-        }
-      })
-    },
-    responses: {
-      200: {
-        description: 'Record was updated'
-      },
-      400: {
-        description:
-          'US State ID and affiliation ID are invalid, roleId is invalid, or status is invalid'
-      },
-      ...responses.unauthed
-    },
-    security: [{ bearerAuth: [] }]
-  }
-};
-
 
 const affiliationRoutes = {
   '/affiliations/me': {
