+++ api/routes/apds/files/post.endpoint.js
@@ -8,60 +8,9 @@ const {
   login,
   unauthenticatedTest,
   unauthorizedTest
-} = require('../../endpoint-tests/utils');
+} = require('../../../endpoint-tests/utils');
 
 describe('APD files endpoints', () => {
-  describe('Get a file associated with an APD | GET /apds/:id/files/:fileID', () => {
-    const db = getDB();
-    beforeAll(() => setupDB(db));
-    afterAll(() => teardownDB(db));
-
-    const url = (apdID, fileID) => `/apds/${apdID}/files/${fileID}`;
-
-    describe('when authenticated as a user with permission', () => {
-      let api;
-
-      beforeAll(async () => {
-        api = login();
-      });
-
-      it('with a non-existant apd ID', async () => {
-        const response = await api.get(
-          url(9000, '74aa0d06-ae6f-472f-8999-6ca0487c494f')
-        );
-
-        expect(response.status).toEqual(400);
-        expect(response.data).toMatchSnapshot();
-      });
-
-      it(`with an APD in a state other than the user's state`, async () => {
-        const response = await api.get(
-          url(4000, '74aa0d06-ae6f-472f-8999-6ca0487c494f')
-        );
-
-        expect(response.status).toEqual(400);
-        expect(response.data).toMatchSnapshot();
-      });
-
-      it('with an APD that is not associated with the file', async () => {
-        const response = await api.get(
-          url(4002, '74aa0d06-ae6f-472f-8999-6ca0487c494f')
-        );
-
-        expect(response.status).toEqual(400);
-        expect(response.data).toMatchSnapshot();
-      });
-
-      it('with a valid request', async () => {
-        const response = await api.get(
-          url(4001, '74aa0d06-ae6f-472f-8999-6ca0487c494f')
-        );
-
-        expect(response.status).toEqual(200);
-        expect(response.data).toMatchSnapshot();
-      });
-    });
-  });
 
   describe('Upload a file associated with an APD | POST /apds/:id/files', () => {
     const db = getDB();
