+++ api/routes/auth/certifications/files/get.js
@@ -1,0 +1,37 @@
+
+const { loggedIn } = require('../../../../middleware/auth');
+const { can } = require('../../../../middleware');
+const logger = require('../../../../logger')('auth certifications post');
+
+const { getFile: get} = require('../../../../files');
+
+module.exports = (
+  app,
+  {
+    getFile = get,
+  } = {}
+) => {
+  logger.silly('setting up GET /auth/certifications/files/:fileID route');
+
+  app.get(
+    '/auth/certifications/files/:fileID',
+    loggedIn,
+    can('view-state-certifications'),
+    async (req, res, next) => {
+      try {
+        const file = await getFile(req.params.fileID);
+        if(file) {
+          res.send(file).end();
+        } else {
+          res.status(400).end();
+        }
+      } catch (e) {
+        logger.error({ id: req.id, message: 'error fetching file' });
+        logger.error({ id: req.id, message: e });
+        next(e);
+      }
+    }
+  );
+
+
+};
