+++ api/routes/auth/certifications/files/index.js
@@ -1,0 +1,11 @@
+const logger = require('../../../../logger')('auth certifications route index');
+const get = require('./get');
+const post = require('./post');
+
+module.exports = (app,
+                  { getEndpoint = get, postEndpoint = post } = {}
+) => {
+  logger.debug('setting up state admin certifications files endpoints');
+  getEndpoint(app);
+  postEndpoint(app);
+};
