+++ api/routes/auth/certifications/files/post.test.js
@@ -1,13 +1,13 @@
 const tap = require('tap');
 const sinon = require('sinon');
 
-const { can } = require('../../../middleware');
-const { loggedIn } = require('../../../middleware/auth');
+const { can } = require('../../../../middleware');
+const { loggedIn } = require('../../../../middleware/auth');
 
-const filesEndpoint = require('./files');
+const postEndpoint = require('./post');
 
-const mockExpress = require('../../../util/mockExpress');
-const mockResponse = require('../../../util/mockResponse');
+const mockExpress = require('../../../../util/mockExpress');
+const mockResponse = require('../../../../util/mockResponse');
 
 let app;
 let res;
@@ -32,7 +32,7 @@ tap.test('state certifications files endpoints', async endpointTest => {
   });
 
   endpointTest.test('setup', async setupTest => {
-    filesEndpoint(app);
+    postEndpoint(app);
     
     setupTest.ok(
       app.post.calledWith(
@@ -45,15 +45,6 @@ tap.test('state certifications files endpoints', async endpointTest => {
       '/auth/certifications/files POST endpoint is setup'
     );
 
-    setupTest.ok(
-      app.get.calledWith(
-        '/auth/certifications/files/:fileID', 
-        loggedIn,
-        can('view-state-certifications'), 
-        sinon.match.func
-      ),
-      '/auth/certifications/files/:fileId GET endpoint is setup'
-    );
   });
   
   endpointTest.test('POST endpoint for uploading a state certification letter/file', async tests => {
@@ -62,7 +53,7 @@ tap.test('state certifications files endpoints', async endpointTest => {
     const req = {};
     
     tests.beforeEach(async () => {
-      filesEndpoint(app, { ...di });
+      postEndpoint(app, { ...di });
       handler = app.post.args.find(
         args => args[0] === '/auth/certifications/files'
       ).pop();
@@ -124,43 +115,5 @@ tap.test('state certifications files endpoints', async endpointTest => {
       );
     });      
   });    
-  
-  endpointTest.test('GET endpoint for fetching a state certification letter/file', async tests => {
-    let handler; 
-    
-    tests.beforeEach(async () => {
-      filesEndpoint(app, { ...di });
-      handler = app.get.args.find(
-        args => args[0] === '/auth/certifications/files/:fileID'
-      ).pop();
-    });
 
-    tests.test('there is an error retrieving the file', async test => {
-      const error = new Error('some error');
-      
-      di.getFile.rejects(error);
-      
-      await handler(
-        { params: { fileID: 'file id' } },
-        res,
-        next
-      );
-
-      test.ok(next.calledWith(error));
-    });
-    
-    tests.test('file is returned successfully', async test => {
-      const file = {};
-      di.getFile.resolves(file);
-      
-      await handler(
-        { params: { fileID: 'file id' } },
-        res,
-        next
-      );
-
-      test.ok(res.send.calledWith(file), 'sends the file');
-      test.ok(res.end.calledAfter(res.send), 'response is terminated');
-    });    
-  });
 });
