+++ api/routes/index.js
@@ -2,27 +2,31 @@ const swaggerUi = require('swagger-ui-express');
 const logger = require('../logger')('routes index');
 const affiliations = require('./affiliations');
 const apds = require('./apds');
-const apdsEvents = require('./apds/events')
+const apdsEvents = require('./apds/events');
+const apdsFiles = require('./apds/files')
 const auth = require('./auth');
 const docs = require('./docs');
 const roles = require('./roles');
 const states = require('./states');
-const statesAffiliations = require('./states/affilitations');
+const stateAffiliations = require('./states/affilitations');
 const users = require('./users');
 const openAPI = require('./openAPI');
 
 module.exports = (
   app,
-  affiliationsEndpoint = affiliations,
-  apdsEndpoint = apds,
-  apdsEventsEndpoint = apdsEvents,
-  authEndpoint = auth,
-  docsEndpoint = docs,
-  rolesEndpoint = roles,
-  statesEndpoint = states,
-  statesAffiliationEndpoint = statesAffiliations,
-  usersEndpoint = users,
-  openAPIdoc = openAPI
+  {
+      affiliationsEndpoint = affiliations,
+      apdsEndpoint = apds,
+      apdsEventsEndpoint = apdsEvents,
+      apdsFilesEndpoint = apdsFiles,
+      authEndpoint = auth,
+      docsEndpoint = docs,
+      rolesEndpoint = roles,
+      statesEndpoint = states,
+      stateAffiliationEndpoint = stateAffiliations,
+      usersEndpoint = users,
+      openAPIdoc = openAPI
+    } = {}
 ) => {
   logger.debug('setting up routes for affilitions');
   affiliationsEndpoint(app);
@@ -30,6 +34,8 @@ module.exports = (
   apdsEndpoint(app);
   logger.debug('setting up routes for apds/events');
   apdsEventsEndpoint(app)
+  logger.debug('setting up routes for apds/files');
+  apdsFilesEndpoint(app)
   logger.debug('setting up routes for auth');
   authEndpoint(app);
   logger.debug('setting up routes for docs');
@@ -39,7 +45,7 @@ module.exports = (
   logger.debug('setting up routes for states');
   statesEndpoint(app);
   logger.debug('setting up routes for states/affiliation');
-  statesAffiliationEndpoint(app)
+  stateAffiliationEndpoint(app)
   logger.debug('setting up routes for users');
   usersEndpoint(app);
