+++ api/routes/index.test.js
@@ -15,25 +15,31 @@ tap.test('endpoint setup', async endpointTest => {
   const affiliationsEndpoint = sinon.spy();
   const apdsEndpoint = sinon.spy();
   const apdsEventsEndpoint = sinon.spy();
+  const apdsFilesEndpoint = sinon.spy();
   const authEndpoint = sinon.spy();
   const docsEndpoint = sinon.spy();
   const meEndpoint = sinon.spy();
   const rolesEndpoint = sinon.spy();
   const statesEndpoint = sinon.spy();
+  const stateAffiliationEndpoint = sinon.spy()
   const usersEndpoint = sinon.spy();
   const openAPI = {};
 
   endpointIndex(
     app,
-    affiliationsEndpoint,
-    apdsEndpoint,
-    apdsEventsEndpoint,
-    authEndpoint,
-    docsEndpoint,
-    rolesEndpoint,
-    statesEndpoint,
-    usersEndpoint,
-    {}
+    {
+      affiliationsEndpoint,
+      apdsEndpoint,
+      apdsEventsEndpoint,
+      apdsFilesEndpoint,
+      authEndpoint,
+      docsEndpoint,
+      rolesEndpoint,
+      statesEndpoint,
+      stateAffiliationEndpoint,
+      usersEndpoint,
+      openAPIdoc:{}
+    }
   );
 
   endpointTest.ok(
@@ -47,6 +53,11 @@ tap.test('endpoint setup', async endpointTest => {
   endpointTest.ok(
     apdsEventsEndpoint.calledWith(app),
     'apds events endpoint is setup with the app'
+  );
+
+  endpointTest.ok(
+    apdsFilesEndpoint.calledWith(app),
+    'apds files endpoint is setup with the app'
   );
 
   endpointTest.ok(
@@ -68,6 +79,10 @@ tap.test('endpoint setup', async endpointTest => {
   endpointTest.ok(
     statesEndpoint.calledWith(app),
     'states endpoint is set up with the app'
+  );
+  endpointTest.ok(
+    stateAffiliationEndpoint.calledWith(app),
+    'state Affiliation endpoint is set up with the app'
   );
   endpointTest.ok(
     usersEndpoint.calledWith(app),
